requirements.txt
flask
numpy
pillow
scikit-image
tensorflow
keras==2.15
mtcnn
opencv-python